Optimize CPU Scheduler and GPU Path Tracing
- CPU Optimization: Replaced `thread::park_timeout(10ms)` with `thread::park()` in `Worker::start` to eliminate busy-waiting and reduce latency.
- GPU Optimization: Implemented Wavefront Path Tracing by splitting the monolithic `ComputePathTracingIndirectPass` into `ComputeRayTraversal` and `ComputeBounceShade` passes. This improves occupancy and reduces divergence.
- Shader Optimization: Reduced BVH traversal stack size from 64 to 32 in `raytracing.inc` to reduce register pressure.
- Cleaned up unused `compute_indirect_bounce.rs` by merging its logic into `compute_pathtracing_indirect.rs`.
