+++ crates/graphics/src/passes/mod.rs
@@ -5,7 +5,6 @@ pub use self::compute_depth_first::*;
 pub use self::compute_depth_pyramid::*;
 pub use self::compute_diffuse_lighting::*;
 pub use self::compute_gbuffer::*;
-pub use self::compute_indirect_bounce::*;
 pub use self::compute_indirect_setup::*;
 pub use self::compute_instances::*;
 pub use self::compute_pathtracing_direct::*;
@@ -24,7 +23,6 @@ pub mod compute_depth_first;
 pub mod compute_depth_pyramid;
 pub mod compute_diffuse_lighting;
 pub mod compute_gbuffer;
-pub mod compute_indirect_bounce;
 pub mod compute_indirect_setup;
 pub mod compute_instances;
 pub mod compute_pathtracing_direct;
